data-visualizer/backend/requirements.txt
FastAPI
pandas
python-binance
uvicorn
python-dotenv